chore: logger wrapper updated
