--- system/console/wrapper/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@web3api/console-wrapper",
-  "description": "Web3API for logging",
-  "private": true,
-  "version": "0.0.1-prealpha.48",
-  "scripts": {
-    "build": "w3 build",
-    "deploy": "w3 build --ipfs http://127.0.0.1:5001 --test-ens console.eth",
-    "test": "w3 query ./recipes/e2e.json",
-    "test:ci": "yarn test:env:up && yarn deploy && yarn test && yarn test:env:down",
-    "test:env:up": "w3 test-env up",
-    "test:env:down": "w3 test-env down"
-  },
-  "devDependencies": {
-    "@web3api/cli": "0.0.1-prealpha.48",
-    "@web3api/wasm-as": "0.0.1-prealpha.48",
-    "assemblyscript": "0.19.1",
-    "js-yaml": "3.14.0"
-  }
-}
