--- system/console/wrapper/src/query/schema.graphql
@@ -1,19 +1,0 @@
-#import { Query } into Logger from "w3://ens/js-logger.web3api.eth"
-
-type Query {
-  debug(
-    message: String!
-  ): Boolean!
-
-  info(
-    message: String!
-  ): Boolean!
-
-  warn(
-    message: String!
-  ): Boolean!
-
-  error(
-    message: String!
-  ): Boolean!
-}
