--- system/console/wrapper/web3api.yaml
@@ -1,9 +1,0 @@
-description: Console log Web3Api
-repository: https://github.com/web3-api/monorepo
-format: 0.0.1-prealpha.1
-query:
-  schema:
-    file: ./src/query/schema.graphql
-  module:
-    language: wasm/assemblyscript
-    file: ./src/query/index.ts
