+++ system/logger/wrapper/.gitignore
@@ -4,8 +4,8 @@ yarn-error.log
 coverage
 .vscode
 *.log
-w3
-.w3
+wrap
+.polywrap
 .DS_Store
 report.*
 .env
