+++ system/logger/wrapper/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@polywrap/console-wrapper",
+  "description": "Polywrap for logging",
+  "private": true,
+  "version": "0.0.1-prealpha.88",
+  "scripts": {
+    "build": "polywrap build",
+    "codegen": "polywrap codegen",
+    "deploy": "polywrap deploy",
+    "test": "polywrap query ./recipes/e2e.json",
+    "test:ci": "yarn test:env:up && yarn test && yarn test:env:down",
+    "test:env:up": "polywrap infra up --modules=eth-ens-ipfs",
+    "test:env:down": "polywrap infra down --modules=eth-ens-ipfs"
+  },
+  "devDependencies": {
+    "@polywrap/logger-plugin-js": "0.0.1-prealpha.88",
+    "@polywrap/wasm-as": "0.0.1-prealpha.88",
+    "polywrap": "0.0.1-prealpha.88",
+    "assemblyscript": "0.19.1",
+    "js-yaml": "3.14.0"
+  }
+}
