+++ system/logger/wrapper/polywrap.deploy.yaml
@@ -1,0 +1,5 @@
+format: 0.0.1-prealpha.1
+stages:
+  ipfs_deploy:
+    package: ipfs
+    uri: fs/./build
