+++ system/logger/wrapper/polywrap.yaml
@@ -1,0 +1,9 @@
+format: 0.0.1-prealpha.9
+name: Logger wrapper
+language: wasm/assemblyscript
+schema: ./src/schema.graphql
+module: ./src/index.ts
+deploy: ./polywrap.deploy.yaml
+import_redirects:
+  - uri: wrap://ens/js-logger.polywrap.eth
+    schema: ./node_modules/@polywrap/logger-plugin-js/build/schema.graphql
