+++ system/logger/wrapper/recipes/e2e.json
@@ -1,6 +1,6 @@
 [
   {
-    "api": "ens/testnet/console.eth",
+    "wrapper": "fs/./build",
     "constants": "./constants.json"
   },
   {
