+++ system/logger/wrapper/recipes/log.graphql
