+++ system/logger/wrapper/src/schema.graphql
@@ -1,0 +1,19 @@
+#import { Module } into Logger from "wrap://ens/js-logger.polywrap.eth"
+
+type Module implements Logger_Module {
+    debug(
+        message: String!
+    ): Boolean!
+
+    info(
+        message: String!
+    ): Boolean!
+
+    warn(
+        message: String!
+    ): Boolean!
+
+    error(
+        message: String!
+    ): Boolean!
+}
